+++ whatsnew.md
@@ -1,5 +1,4 @@
 # This file will provide extra info on new updates / commits.
 
 # WHAT'S NEW [23/12/2025]
-  - Added PEB extraction (PEB, Debugged, ImagePath, CommandLine)
-  - Added an output file to show correctness. Since most of TEB & PEB extraction are current-process, it will **only** show the output of our process, ***MrtTInfoTest.exe***. If in the future, I (or others) add other-process reading, it will show other process's output as well. 
+  - Added PEB Module Loader structs